pyproject.toml
[project]
name = "mlops-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "matplotlib>=3.10.6",
    "mlflow>=3.4.0",
    "openai-agents>=0.3.3",
]
